pyproject.toml
[project]
name = "self-healing-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
]
